requirements.txt
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
openai>=1.1.1
tiktoken>=0.5.1
tqdm>=4.66.0
streamlit>=1.28.0
